backend/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
google-cloud-storage>=2.10.0
google-generativeai>=0.3.0
Pillow>=10.0.0
fpdf2>=2.7.0
pdfplumber>=0.10.0
yake>=0.4.8
python-dateutil>=2.8.2
flask-cors==4.0.0
firebase-admin>=6.4.0
